load client state from server

+++ utils.py
@@ -45,7 +45,7 @@ def save_state(path,name):
     '''
     
     if not os.path.isdir(path):
-        utils.create_directory(path) 
+        create_directory(path) 
     filename = path + "/" + name
     
     bpy.ops.wm.collada_export(filepath=filename)
@@ -57,13 +57,50 @@ def load_state(path,name):
     path         -- a string that contains the filepath of the folder where the file will be loaded
     name         -- a string that contains the filename of the collada file to load
     '''
-    
-    #clear the scene to remove objects that are not part of the state to load
-    bpy.ops.object.select_all(action='SELECT')
-    bpy.ops.object.delete()
-    
-    try:
-        filename = path + "/" + name + ".dae"
+    filename = path + "/" + name + ".dae"
+    if os.path.isfile(filename):
+        #clear the scene to remove objects that are not part of the state to load
+        bpy.ops.object.select_all(action='SELECT')
+        bpy.ops.object.delete()
         bpy.ops.wm.collada_import(filepath=filename)
-    except IOError:
-        pass
+        
+def check_file(path,name):
+    '''checks if a file exists
+    
+    Parameters
+    path        -- a string containing the filepath of the folder where the file is located
+    name        -- a string containing the filename of the file to check
+    
+    Return Value
+    exist_flag  -- a boolean value used to indicate a file's existence 
+                -- True --> the file exists
+                -- False --> the file does not exist
+                
+    '''
+    
+    filename = path + "/" + name + ".dae"
+    if os.path.isfile(filename):
+        exist_flag = True
+    elif not os.path.isfile(filename):
+        exist_flag = False
+        
+    return exist_flag
+
+def check_dir(path):
+    '''checks if a directory exists
+    
+    Parameters
+    path        -- a string containing the filepath of the directory to check
+    
+    Return Value
+    exist_flag    -- a boolean value used to indicate a directory's existence
+    
+    '''
+    
+    if os.path.isdir(path):
+        exist_flag = True
+    elif not os.path.isdir(path):
+        exist_flag = False
+        
+    return exist_flag
+
